+++ ChangeLog.md
@@ -2,6 +2,23 @@
 
 ## 2024/02/05
 
+### WizardWrx.Core 9.0.357
+
+Correct a logic error in the `ExtractBetweenIndexOfs` string extension method
+overload that takes into account the length of the substring that marks the
+left boundary of the substring to extract by adding its length to the specified
+left boundary position. 
+
+The eror took the form of a nonsensical test that caused it to assume that the
+bounds were invalid. The objective of the invalid test was to prevent the native
+SubString method from throwing because of one or more invalid charaacter
+positions.
+
+This correction resolves the puzzling issue that prompted the harmless change
+introduced into the version of ExtractBoundedSubstrings in build 9.0.355.
+
+## 2024/02/05
+
 ### WizardWrx.AssemblyUtils 9.0.275
 
 Implement two new methods, `GetAssemblyGUID` and `GetAssemblyTargetFramework`,